refactor: migrate data function to sheet.ts
